third_party/WebKit/LayoutTests/wake_lock/wakelock-api.html

<!DOCTYPE html>
<html>
<body>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>

test(function() {
    assert_idl_attribute(screen, 'keepAwake');
    assert_equals(typeof screen.keepAwake, 'boolean');
}, 'Test that the WakeLock API is present.');

test(function() {
    screen.keepAwake = true;
    assert_true(screen.keepAwake);

    screen.keepAwake = false;
    assert_false(screen.keepAwake);
}, 'Test that screen.keepAwake preserves set value');

</script>
</body>
</html>
